Description:修改GetAllDisplayPhysicalResolution接口返回值、安全告警处理 IssueNo:https://gitee.com/openharmony/window_window_manager/issues/IBJ3LN Binary Source:No Signed-off-by: songjiawei9 <songjiawei9@h-partners.com>

+++ window_scene/screen_session_manager/src/fold_screen_controller/secondary_display_fold_policy.cpp
@@ -347,7 +347,7 @@ void SecondaryDisplayFoldPolicy::SetMainScreenRegion(DMRect& mainScreenRegion)
         mainScreenRegion.height_ = screenParams_[MAIN_STATUS_WIDTH];
     } else if (currentDisplayMode_ == FoldDisplayMode::FULL) {
         mainScreenRegion.posX_ = 0;
-        mainScreenRegion.posY_ = screenParams_[FULL_STATUS_OFFSET_X];
+        mainScreenRegion.posY_ = static_cast<int>(screenParams_[FULL_STATUS_OFFSET_X]);
         mainScreenRegion.width_ = screenParams_[SCREEN_HEIGHT];
         mainScreenRegion.height_ = screenParams_[FULL_STATUS_WIDTH];
     } else if (currentDisplayMode_ == FoldDisplayMode::GLOBAL_FULL) {
